bug fix: convert to double instead of integer in DoubleProperty

+++ structr-core/src/main/java/org/structr/core/property/DoubleProperty.java
@@ -106,7 +106,7 @@ public class DoubleProperty extends AbstractPrimitiveProperty<Double> {
 			}
 			
 			if (value instanceof Number) {
-				return ((Number)value).intValue();
+				return ((Number)value).doubleValue();
 			}
 			
 			try {
